Massive refactoring for better separation of concerns and ease of test in all modules
