+++ api/src/document-inteligence/document-inteligence.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { DocumentInteligenceService } from './document-inteligence.service';
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [DocumentInteligenceService],
+  exports: [DocumentInteligenceService],
+})
+export class DocumentInteligenceModule {}
