--- api/src/product/product.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Product } from './product.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Product])],
-  controllers: [],
-  providers: [],
-})
-export class ProductModule {}
