+++ api/src/purchase/services/purchase.service.spec.ts
@@ -1,0 +1,7 @@
+describe('teste', () => {
+  describe('do something', () => {
+    it('should be defined', () => {
+      expect(true).toBeDefined();
+    });
+  });
+});
