+++ api/src/shared/types/documentReceipt.ts
